Created installer script for windows
There is a NSIS script which creates a installer executable in the
build/ directory, upon running puts the tinytcp.exe file into "Program
Files\" under "tinytcp" and adds it to PATH such that you can run
`tinytcp` in the terminal now. And running two in the same terminal
doesn't result in issues.