config.toml file feature added
Needed to adjust the compile scripts in run.sh, run.bat, and flake.nix
to get it to recognize the header file. The toml.hpp file was open
source, I didn't write all that. This works on both windows and linux,
on putting it in ~/.config/tinytcp/config.toml or
~/AppData/Roaming/tinytcp/config.toml depending on OS