Updated version, added examples, fixed some code quality
Version perviously was still template v0.0.1-beta in many places, this
has been updated to v0.4.0 as the previous iteration of this project
never surpassed v0.3.1 and this is a big improvemenet. Two examples
files have been added to the examples/ directory to demonstrate how to
use the library. Code quality was also improved, mainly just converting
variables to const.