Added file sending ability
main.cpp was updated to set message to be the contents of a file if it's
the name of a file when parsing. there is an example.txt file that you
can send to test if it works. README.md has been updated to include an
example of sending a file, as well as the verbose help menu in args.cpp